unsinged int changed to unsigned uint8_t

+++ VM/srcs/introduce_plrs.c
@@ -6,7 +6,7 @@
 /*   By: aelphias <aelphias@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/05 17:09:50 by aelphias          #+#    #+#             */
-/*   Updated: 2021/01/05 17:11:35 by aelphias         ###   ########.fr       */
+/*   Updated: 2021/01/05 21:11:41 by aelphias         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	introduce_plrs(t_plr *plr)
 {
-	unsigned int i;
+	uint8_t i;
 
 	i = 1;
 	ft_printf("Introducting contestants...\n");
